moved the api gateway test to the api gateway service folder and set up a minimal CI workflow to test the service
