+++ .github/workflows/test_api_gateway.yml
@@ -1,0 +1,31 @@
+# .github/workflows/api-gateway-ci.yml
+
+name: API Gateway CI
+
+on:
+  pull_request:
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r services/api-gateway/requirements.txt
+          pip install pytest pytest-cov
+
+      - name: Run tests with coverage
+        run: |
+          python -m pytest tests/unit/test_api_gateway.py -v \
+          --cov=services/api-gateway \
+          --cov-report=term-missing
